Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main()
 {
-    int a,b,c,average;
-    scanf("%d %d %d", &a,&b,&c);
+    float a,b,c,average;
+    scanf("%f %f %f", &a,&b,&c);
     average= (a+b+c)/3;
-    printf("Average: %d\n",average);
+    printf("Average: %f\n",average);
     return 0;
 }
